refactor(cache): centralize image cache wipe in SessionManager.clearSession
The image cache had to be cleared anywhere a session ended (logout in
ProfileViewModel, password change + account delete in PrivacyViewModel).
Three call sites for one invariant — easy to forget when a fourth
logout path lands.

Move ImageCacheCleaner to core, inject it into SessionManager, and call
clear() inside clearSession(). Drop the per-viewmodel wiring. The
version-bump path in AppUpdateMigrator still uses ImageCacheCleaner
directly (it's not a session boundary).
