+++ mobile/app-ui/src/androidMain/kotlin/com/poziomki/app/ui/cache/AndroidImageCacheCleaner.kt
@@ -2,6 +2,7 @@ package com.poziomki.app.ui.cache
 
 import android.content.Context
 import coil3.imageLoader
+import com.poziomki.app.cache.ImageCacheCleaner
 
 class AndroidImageCacheCleaner(
     private val context: Context,
