+++ mobile/app-ui/src/commonMain/kotlin/com/poziomki/app/ui/feature/profile/PrivacyViewModel.kt
@@ -8,7 +8,6 @@ import com.poziomki.app.network.ApiResult
 import com.poziomki.app.network.ApiService
 import com.poziomki.app.session.AppPreferences
 import com.poziomki.app.session.SessionManager
-import com.poziomki.app.ui.cache.ImageCacheCleaner
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -36,7 +35,6 @@ class PrivacyViewModel(
     private val cacheManager: CacheManager,
     private val settingsRepository: SettingsRepository,
     private val appPreferences: AppPreferences,
-    private val imageCacheCleaner: ImageCacheCleaner,
 ) : ViewModel() {
     private val _state = MutableStateFlow(PrivacyState())
     val state: StateFlow<PrivacyState> = _state.asStateFlow()
@@ -168,7 +166,6 @@ class PrivacyViewModel(
             when (apiService.changePassword(currentPassword, newPassword)) {
                 is ApiResult.Success -> {
                     cacheManager.clearAll()
-                    runCatching { imageCacheCleaner.clear() }
                     sessionManager.clearSession()
                     _state.value =
                         _state.value.copy(
@@ -199,7 +196,6 @@ class PrivacyViewModel(
             when (apiService.deleteAccount(password)) {
                 is ApiResult.Success -> {
                     cacheManager.clearAll()
-                    runCatching { imageCacheCleaner.clear() }
                     sessionManager.clearSession()
                     _state.value = _state.value.copy(isDeleting = false)
                     onDeleted()
