+++ mobile/app-ui/src/iosMain/kotlin/com/poziomki/app/Koin.kt
@@ -1,9 +1,9 @@
 package com.poziomki.app
 
+import com.poziomki.app.cache.ImageCacheCleaner
+import com.poziomki.app.cache.NoopImageCacheCleaner
 import com.poziomki.app.di.appModule
 import com.poziomki.app.di.startAppKoin
-import com.poziomki.app.ui.cache.ImageCacheCleaner
-import com.poziomki.app.ui.cache.NoopImageCacheCleaner
 import org.koin.dsl.module
 
 fun initKoin() {
